#1752 moved processing-steps-meta to root
